LPS-60972 The Maven plugin configuration is specific to Liferay
